general updates

+++ otashu-android/src/com/andrewsummers/otashu/activity/ApprenticeEmotionTestActivity.java
@@ -120,7 +120,7 @@ public class ApprenticeEmotionTestActivity extends Activity implements OnClickLi
 
         // get data from bundle
         Bundle bundle = getIntent().getExtras();
-        emotionFocusId = bundle.getInt("emotion_focus_id");
+        emotionFocusId = bundle.getLong("emotion_focus_id");
 
         try {
             // add listeners to buttons
