+++ otashu-android/src/com/andrewsummers/otashu/activity/ApprenticeScaleTestActivity.java
@@ -92,7 +92,7 @@ public class ApprenticeScaleTestActivity extends Activity implements OnClickList
 
         // get data from bundle
         Bundle bundle = getIntent().getExtras();
-        emotionFocusId = bundle.getInt("emotion_focus_id");
+        emotionFocusId = bundle.getLong("emotion_focus_id");
 
         try {
             // add listeners to buttons
